moved code a bit around, cleaned up some classes

+++ src/FbTk/TextButton.cc
@@ -150,7 +150,7 @@ void TextButton::renderForeground(FbWindow &win, FbDrawable &drawable) {
 }
 
 void TextButton::drawText(int x_offset, int y_offset, FbDrawable *drawable) {
-    size_t textlen = text().size();
+    unsigned int textlen = text().size();
     // do text alignment
 
     unsigned int textw = width(), texth = height();
